some documentation added (WIP), added custom log in settings

+++ src/enumy/src/enums_structs_io.rs
@@ -32,7 +32,7 @@ pub struct LogRozpakowywanie {
     pub kontrola_pliku: (u64,Option<u64>),
     pub rozpakowanie: (u32,Option<u32>),
     pub dekompresja: u64,
-    pub deszyfrowanie: (u32,Option<u32>),
+    pub deszyfrowanie: (u64,Option<u64>),
     pub czas: String,
     pub błąd: String,
 }
